Created TextInputComponent
