+++ src/components/addTaskComponent/addTaskComponent.js
@@ -1,34 +1,17 @@
 import React, { useState } from "react";
 import { TasklistComponent } from "./../tasklistComponent/tasklistComponent.js";
+import { TextInputComponent } from "./../textInputComponent/textInputComponent.js";
 
 export const AddTaskComponent = () => {
-  const temporaryTodoList = [];
   const [todoList, setTodoList] = useState([]);
-  const [newTodo, setNewTodo] = useState("");
 
-  function changeHandler(event) {
-    let todoInput = event.target.value;
-    setNewTodo(todoInput);
-  }
-
-  function addTodo() {
-    temporaryTodoList.push(newTodo);
-    setTodoList(temporaryTodoList);
+  function addTodo(newItem) {
+    setTodoList((bestehendesArray) => [...bestehendesArray, newItem]);
   }
 
   return (
     <div>
-      <form>
-        <input
-          type="text"
-          placeholder="Tippe etwas..."
-          value={newTodo}
-          onChange={changeHandler}
-        />
-        <button type="submit" onClick={addTodo}>
-          Hinzufügen
-        </button>
-      </form>
+      <TextInputComponent funktionVonElternKomponente={addTodo} />
       <TasklistComponent toDos={todoList} />
     </div>
   );
